# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Weather API Documentation",
        "description": "API for managing users and consulting weather data from WeatherAPI",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/weather/{city}": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Get current weather for a city",
                "description": "Returns current weather data for the specified city. Also logs the search to user's history.",
                "operationId": "getCurrentWeatherForCity",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "description": "Name of the city",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WeatherResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "City not found or API error"
                    },
                    "422": {
                        "description": "Validation error (e.g., city name too long)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/weather/search": {
            "get": {
                "tags": [
                    "Weather"
                ],
                "summary": "Search for weather locations",
                "description": "Returns a list of locations matching the search query.",
                "operationId": "searchWeatherLocations",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query (e.g., city name fragment)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LocationResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "No locations found or API error"
                    },
                    "422": {
                        "description": "Validation error (e.g., query too short)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "LocationResource": {
                "title": "Location Resource",
                "description": "Location data from weather API search",
                "properties": {
                    "id": {
                        "description": "Unique ID of the location from the weather API",
                        "type": "integer",
                        "readOnly": "true",
                        "example": 2801268
                    },
                    "name": {
                        "description": "Full name of the location",
                        "type": "string",
                        "example": "London, City of London, Greater London, United Kingdom"
                    },
                    "region": {
                        "description": "Region of the location",
                        "type": "string",
                        "example": "City of London, Greater London"
                    },
                    "country": {
                        "description": "Country of the location",
                        "type": "string",
                        "example": "United Kingdom"
                    },
                    "latitude": {
                        "description": "Latitude of the location",
                        "type": "number",
                        "format": "float",
                        "example": 51.52
                    },
                    "longitude": {
                        "description": "Longitude of the location",
                        "type": "number",
                        "format": "float",
                        "example": -0.11
                    }
                },
                "type": "object"
            },
            "WeatherResource": {
                "title": "Weather Resource",
                "description": "Weather data for a specific city",
                "properties": {
                    "city": {
                        "type": "string",
                        "example": "London"
                    },
                    "temperature": {
                        "type": "string",
                        "example": "15.0 °C"
                    },
                    "condition": {
                        "type": "string",
                        "example": "Partly cloudy"
                    },
                    "wind": {
                        "type": "string",
                        "example": "10.0 kph"
                    },
                    "humidity": {
                        "type": "string",
                        "example": "70%"
                    },
                    "local_time": {
                        "type": "string",
                        "example": "2023-10-27 10:00"
                    },
                    "retrieved_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-27T10:05:00.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}